Rh 24

+++ app/src/main/java/com/example/recruiterhunter/domain/repository/VacancyDetailsRepository.kt
@@ -3,5 +3,5 @@ package domain.repository
 import domain.model.vacancy.details.Vacancy
 
 interface VacancyDetailsRepository {
-    suspend fun fetchDetails(vacancyId: Int): Result<Vacancy>
+    suspend fun fetchDetails(vacancyId: Long): Result<Vacancy>
 }
